Decide to separate calling oracle from token funcs
1. Basically, no matter whether users call indirectly using
transferAndCall or directly calling a function hosting the oracle
operations, they will consume msg.sender's gas since a smart contract
does not possess its own eth.
2. Also, something recorded here in case of future usage: after you
create a new contract, you better wait for at least 5 mins before
calling transferAndCall to it. If this contract is successfully
registered by many minners, it will be reflected by the MetaMask (which
will show its estimated gas, if it is not showing, which means that
the smart contract might not be identified yet or its codes consude too
much gas). Plus, you can also try to adjust the _value parameter in the
transferAndCall, which might makes it works.